JIT: Refactor physical promotion in preparation of generalized assignment decomposition
This refactors the physical promotion to make the generalized assignment
decomposition PR more manageable to review.

* Pull Replacement and ReplaceVisitor into the header
* Introduce promotiondecomposition.cpp that deals with handling
  assignments
* Rename StatementList -> DecompositionStatementList and unnest it

Also adds some more JITDUMPs when picking promotions, but otherwise
should not affect behavior.
